test/transforms/liftings/pointcloud2simplicial/test_alpha_complex_lifting.py
"""Test the alpha complex lifting."""

import torch

from modules.data.utils.utils import load_manual_points
from modules.transforms.liftings.pointcloud2simplicial.alpha_complex_lifting import (
    AlphaComplexLifting,
)


class TestAlphaComplexLifting:
    """Test the AlphaComplexLifting class."""

    def setup_method(self):
        # Load the graph
        self.data = load_manual_points()

        # Initialise the AlphaComplexLifting class
        self.lifting = AlphaComplexLifting(complex_dim=3, alpha=25.0)

    def test_lift_topology(self):
        """Test the lift_topology method."""

        # Test the lift_topology method
        lifted_data = self.lifting.forward(self.data.clone())

        expected_incidence_1 = torch.tensor(
            [
                [1.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
                [1.0, 0.0, 1.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0],
                [0.0, 1.0, 1.0, 0.0, 1.0, 1.0, 0.0, 0.0, 0.0],
                [0.0, 0.0, 0.0, 1.0, 1.0, 0.0, 0.0, 0.0, 0.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 0.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 1.0],
                [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 1.0],
            ]
        )

        assert (
            expected_incidence_1 == lifted_data.incidence_1.to_dense()
        ).all(), "Something is wrong with incidence_1 (nodes to edges)."

        expected_incidence_2 = torch.tensor(
            [
                [1.0, 0.0, 0.0],
                [1.0, 0.0, 0.0],
                [1.0, 1.0, 0.0],
                [0.0, 1.0, 0.0],
                [0.0, 1.0, 0.0],
                [0.0, 0.0, 0.0],
                [0.0, 0.0, 1.0],
                [0.0, 0.0, 1.0],
                [0.0, 0.0, 1.0],
            ]
        )

        assert (
            abs(expected_incidence_2) == lifted_data.incidence_2.to_dense()
        ).all(), "Something is wrong with incidence_2 (edges to triangles)."

        expected_incidence_3 = torch.tensor([])

        assert (
            abs(expected_incidence_3) == lifted_data.incidence_3.to_dense()
        ).all(), "Something is wrong with incidence_3 (triangles to tetrahedrons)."
